change score from integer to double

+++ database/migrations/2019_11_28_183430_add_score_column_to_books_table.php
@@ -14,7 +14,7 @@ class AddScoreColumnToBooksTable extends Migration
     public function up()
     {
         Schema::table('books', function (Blueprint $table) {
-            $table->integer('score');
+            $table->double('score', 8, 2);
         });
     }
